Change hdg from Double to Integer in HfpJson

+++ src/main/java/fi/hsl/common/hfp/HfpJson.java
@@ -39,7 +39,7 @@ public class HfpJson {
 
         public Double spd;
 
-        public Double hdg;
+        public Integer hdg;
 
         public Double lat;
